requirements.txt
accelerate==0.30.0
diffusers==0.27.2
transformers==4.40.2
safetensors==0.4.3
einops==0.7.0
lpips==0.1.4 
cmake==3.25.0
gradio==3.47.1
matplotlib==3.7.1
numpy==1.24.1
opencv-python==4.8.0.76
pandas==2.0.2
Pillow==9.3.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.65.0
xformers==0.0.20
yacs==0.1.8